Remove libXMI - Part 1 of 4: add our own polygon filler.
Enblend depends on libXMI, an ancient graphics library.  It uses only
one function, the polygon filler that converts a blend region defined
by a border to a blend mask.  This series (1-4) of patches will add a
new, home-grown polygon filler that only depends on Vigra.  At the
end of the series the dependency on libXMI will be removed.