Simplify our OpenCL interface.
It is much easier to pass around a cl::Context pointer, which even
works without having the class definition of cl::Context available.
The context gets associated with the preferred platform and the
preferred device is the first one in the list of all available GPU
devices on the platform.  To recover the device list use, e.g.
        std::vector<cl::Device> devices =
            GPUContext->getInfo<CL_CONTEXT_DEVICES>();