Parameterize a highlight optimizer bracketing constant and catch a related exception.
        Pull out the maximum number of tries to bracket the lightness
(J) of an out-of-cube highlight pixel (before the 1d-optimizer takes
over) in the new command-line parameter
`maximum_highlight_bracket_tries'.  The old default was 100 tries, the
new default is 1000.

        If we failed to bracket the highlight pixel's minimum along
the lightness axis we threw a runtime error and aborted.  Now we throw
a tailored exception, catch it in a narrow enclosing scope, and fall
back to the initial guess of J, because we truly have no better idea
at this point of computation.  As in any code path, out-of-cube RGB
components will be mopped up by the final min-max limiter, though in
an un-photographic way.  Compiling with
        DEBUG_SHADOW_HIGHLIGHT_STATISTICS
defined will reveal these cases.

This is a transplant of 0f3c8eea75a2.