Add a heuristic to cope with shadows that disguise as highlights.
For quite some time we have been separating shadows and highlights
that need fixing according to which plane of the RGB-cube they are
next to.  The six sides of the cube define two sets of planes through
the connection with pure black (0, 0, 0) or pure white (1, 1, 1).

With newer versions of LittleCMS sometimes shadows, which seem to need
fixing disguise as highlights, i.e. no RGB component is negative, but
at least one is larger than one.  However the J-component of the pixel
is very small, characterizing it as shadow.  A heuristic constant,
which is controlled by parameter
        highlight-disguised-as-shadow-lightness
determines a "J black point" below which these highlights in disguise
are mapped to pure black.

This patch cleans up seemingly random, bright-colored pixels that
sometimes show up when fusing images with high contrast and "large"
profiles.

This is a transplant of 2d06e85c0487.