Change handling of empty `--parameter' options.
(1) Empty '--parameter' values set the value "1".  This is similar to the
    C preprocessor, where
            #define FOO
    defines `FOO' and assigns the integer value 1 to it.
(2) Expand the valid values of boolean parameters to "0" and "1"
    for `false' and `true'.

Together, (1) and (2) allow for a shorter syntax of "switch-on"
parameters, e.g.
        --parameter=debug-dijkstra
instead of the old
        --parameter=debug-dijkstra=true
which is still valid.  Hungry assignment operators will be rejected,
i.e.
        --parameter=debug-dijkstra=
triggers an error.