Abstain from using tiff2ps(1); use convert(1) instead.
Rationale: When used with the `eps3'-output tag, convert(1) produces
EPS files approximately as compact as those generated with tiff2ps(1)
running under option `-3'.  Prefer convert(1) because we are already
using it for other conversions and tiff2ps(1) only is part of the
somewhat obscure TIFF-utilities.

Hat tip to our trusty Debian maintainer, Andreas Metzler, for the
`eps3'-tag trick.