Fix wrong usage of `std::unique_ptr' vs `new []' in message handler modules.
The correct template parameter for `std::unique_ptr' is `char []' if
the object is allocated with operator `new []'.  In our case it is
easier to resize the message and overwrite its contents using a
pointer from `std::string::operator[]'.  The latter is valid use of
`operator[]' and guaranteed to work by C++11.

This is "0002-Fix-wrong-usage-of-boost-scoped_ptr-new-vs-new.patch" in
https://bugs.launchpad.net/enblend/+bug/1332323