Shoo away undeserved warning about incompatible colorspaces.
Previously we warned about possible problems with different
colorspaces if a pair of input images came with ICC profiles that did
not match bytewise.  This was prudent.  Darktable
(http://www.darktable.org/) attaches a freshly generated `REC 709' ICC
profile to each image.  Thus, the profiles' creation times disagree
and Enblend and Enfuse warn about different profiles although it is
only divergent meta-data.  This patch relaxes the test to equality of
the profiles' descriptions and fixes issue #1364695
    https://bugs.launchpad.net/enblend/+bug/1364695

This is a backport of change #3e1c4bc71018.