Fix longstanding bug in primary seam line vectorization (1 of 2). Part 1: Fix possible out-of-bounds access in function minCostPath().
Function minCostPath() accesses the cost image (as well as in turn
image `costSoFar') at `endingPoint' without a test whether the point
belongs to the image.  If `endingPoint' lies outside of the cost image
the access triggers a seg-fault.  (No such problem occurs with
`startingPoint'.)

The patch introduces a test for the location of `endingPoint' and
skips the point if is outside of the valid region.  We need the same
kind of test in the seam-line visualization code in "postoptimizer.h".

Transplant of 255771c2e58b