Fix longstanding bug in primary seam line vectorization (2 of 2). Part 2: Fix rare off-by-one error in mask vectorization code.
The problem originates from a plain integer division which truncates
(as demanded by the C Standard).  However, we need round-to-nearest
semantics.  In theory, almost every coarse-mask generation is
affected, never a fine mask, though.

This should fix issue #1356551
        https://bugs.launchpad.net/enblend/+bug/1356551
and looks like a good candidate for back-porting to the Stable Branch.

Patches (1 of 2) and (2 of 2) were both reviewed by Rosomack.

Transplant of 95c7e90f2be8