Improve error messages for long options that require arguments.
If a long option that requires qrguments was lacking its arguments the
error handler of short options was triggerd and overwhelmed by the
error code only printed a (confusing) hex value.

This patch installs a separate handler for missing arguments of long
options and tryes to reconstruct the original option name for a more
helpful message.

Backport of change 11a20ae8f7be.