Allow grayscale images to bring their own grayscale profiles.
En passant fix a little inconsistency between Enblend and Enfuse:
prefer the white point of the input profile and only fall back to D50
if it is unusable.  (We need the white point when we set up the color
transform to LAB space.)

This is a transplant of change #12b596d6fb50.