Make option `--blend-colorspace' available in the Stable Branch. The functionality remains the same, though (Stable Branch, remember?).
Change a92ec45d7566 (in the Development Branch) introduces L*a*b* and
L*u*v* color spaces for pyramidal blending operations in addition to
the existing ones, RGB and CIECAM.  For the Stable Branch we only pick
the new command-line option `--blend-colorspace'.  This will allow the
users to switch from `--ciecam' to `--blend-colorspace=ciecam' and
from `--no-ciecam' to `--blend-colorspace=identity' and eases an
unpgrade to the next version of Enblend/Enfuse, where `--ciecam' and
`--no-ciecam' will be deprecated.