Allow to consider the metadata of any of the input images not just the first input image.
The default remains to use the metadata of the first image with valid
metadata on the commandline.  Parameter `metadata-source-image-index'
allows to override the index with a C-style index starting at zero.

The current implementation does not couple to layered input images
well enough.

The completely brain-dead utilization of `std::auto_ptr', which have
long been deprecated and even longer have been known for being
broken-as-designed in the Exiv2 library forces us to use placement-new
and similar C++ magic to get an array of metadata that corresponds to
our array of input images.  Where is the Exiv2 fork that uses
`std::unique_ptr'?